Set up auxiliary models

Set up an auxiliary model from that differs from the target model.

Usage

auxiliary_model(
  time_varying_variables,
  time_invariant_variables,
  linear,
  heterogeneity,
  use_open_mx,
  verbose,
  ...
)

Arguments

time_varying_variables

List of character vectors containing names of the time-varying variables. The number entries in the list corresponds to the number of univariate time-series. Each character vector contains the time-ordered variable names of a univariate time-series starting with the first measurement occasion.

time_invariant_variables

List of character vectors containing names of the time-invariant variables. List must have the same length as list in argument time_varying_variables.

linear

Logical (TRUE = default / FALSE) indicating if the model is linear in observed variables (TRUE).

heterogeneity

Character vector indicating the type of unobserved heterogeneity. Admissible values are "homogeneous", "additive", "autoregressive", and "cross-lagged" (or any non-conflicting combination).

use_open_mx

Logical (TRUE / FALSE default) indicating if lavaan (FALSE) or OpenMx (TRUE) should be used.

verbose

Integer number describing the verbosity of console output. Admissible values: 0: no output (default), 1: user messages, 2: debugging-relevant messages.

...

not used

Value

An object of class panelSEM.

References

Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal Effects of Interventions versus Predicting Future Outcomes, Structural Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492, DOI: 10.1080/10705511.2020.1780598