Create Empty Internal List

Creates an empty internal list of predefined structure. The list will subsequently be filled while the function fit_panel_sem is running.

Usage

create_empty_list(verbose)

Arguments

verbose

Integer number describing verbosity of console output. 0...no output (default); 1...user messages; 2...debugging-relevant messages.

Value

An empty internal list with the following structure:

List of XX	$info_data: List of XX

..$data: data.frame

User-specified data set.	..$data_product_terms: data.frame

Data set with product terms of observed variables.	..$n_obs: int(0)

Number of observations.	..$n_var: int(0)

Total number of variables in the data set.	..$var_names: chr(0)

Names of variables.	$info_model: List of XX

..$n_ocassions: int(0)

Number of measurement occasions.	..$n_processes: int(0)

Number of dynamic processes.	..$n_time_invariant int(0)

Number of observed time-invariant variables	..$linear logical(0)

TRUE if the model is linear in observed variables.	..$heterogeneity chr(0)

Specify type of unobserved heterogeneity.	..$use_resamples logical(0)

. If TRUE, a resampling procedure for starting values and model diagnostics is performed.	..$use_open_mx logical(0)

If TRUE, OpenMx is used, otherwise lavaan is used.	$info_variables: List of XX

..$user_names_time_varying: char[0,0]

User-Specified names of observed time-varying variables.	..$generic_names_time_varying: char[0,0]

Generic names of observed time-varying variables.	..$names_processes: char[0,0]

Names of dynamic processes.	..$info_time_invariant_varianbles: char[0,0]

Names and causal structure of observed time-invariant variables.	..$names_time_invariant_unique: char[0,0]

Names of unique observed time-invariant variables.	..$names_time_invariant_unobserved_additive: char[0,0]

Names of UNobserved time-invariant variables.	..$names_time_invariant_unobserved_autoregressive: char[0,0]

Names of UNobserved time-invariant variables.	..$names_time_invariant_unobserved_cross_lagged: char[0,0]

Names of UNobserved time-invariant variables.	$info_parameters: List of XX

..$C_table: data.frame

Table with labeled model parameters (structural coefficients).	..$Psi_table: data.frame

Table with labeled model parameters (covariance matrix).	model_matrices: List of XX

..$C_labels: char[0,0]

Character matrix of labels of structural coefficients.	..$Psi_labels: char[0,0]

Character matrix of labels of covariance matrix.	..$select_observed_only: num[0,0]

Numeric matrix to select observed variables.	..$C_start: num[0,0]

Numeric matrix of starting values of structural coefficients.	..$Psi_start: num[0,0]

Numeric matrix of starting values of covariance matrix.	$model_syntax: List of XX

..$lavaan: list

Model specified in lavaan.	..$OpenMx: list

Model specified in OpenMx.	$control: List of XX

..$verbose: int(0)

References

Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal Effects of Interventions versus Predicting Future Outcomes, Structural Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492, DOI: 10.1080/10705511.2020.1780598